+++ examples/sharing.rs
@@ -8,21 +8,23 @@ use support::gl;
 use glutin::GlContext;
 
 fn main() {
-    let mut events_loop = glutin::EventsLoop::new();
+    let mut el = glutin::EventsLoop::new();
     let mut size = glutin::dpi::PhysicalSize::new(768., 480.);
 
-    let context = glutin::ContextBuilder::new();
-    let gl_context = glutin::Context::new(&events_loop, context).unwrap();
+    let headless_context = glutin::ContextBuilder::new()
+        .build_headless(&el)
+        .unwrap();
 
-    let window = glutin::WindowBuilder::new().with_title("A fantastic window!")
+    let wb = glutin::WindowBuilder::new().with_title("A fantastic window!")
         .with_dimensions(glutin::dpi::LogicalSize::from_physical(size, 1.0));
-    let context = glutin::ContextBuilder::new()
-        .with_shared_lists(&gl_context);
-    let gl_window = glutin::GlWindow::new(window, context, &events_loop).unwrap();
+    let combined_context = glutin::ContextBuilder::new()
+        .with_shared_lists(&headless_context)
+        .build_combined(wb, &el)
+        .unwrap();
 
-    let _ = unsafe { gl_window.make_current() };
-    println!("Pixel format of the window's GL context: {:?}", gl_window.get_pixel_format());
-    let glw = support::load(&gl_window.context());
+    let _ = unsafe { combined_context.make_current() };
+    println!("Pixel format of the window's GL context: {:?}", combined_context.get_pixel_format());
+    let glw = support::load(&combined_context.context());
 
     let mut render_tex = 0;
     unsafe {
@@ -51,8 +53,8 @@ fn main() {
         );
     }
 
-    let _ = unsafe { gl_context.make_current() };
-    let glc = support::load(&gl_context);
+    let _ = unsafe { headless_context.make_current() };
+    let glc = support::load(&headless_context);
 
     let mut context_fb = 0;
     unsafe {
@@ -70,19 +72,19 @@ fn main() {
 
     let mut running = true;
     while running {
-        events_loop.poll_events(|event| {
+        el.poll_events(|event| {
             println!("{:?}", event);
             match event {
                 glutin::Event::WindowEvent { event, .. } => match event {
                     glutin::WindowEvent::CloseRequested => running = false,
                     glutin::WindowEvent::Resized(logical_size) => {
-                        let _ = unsafe { gl_window.make_current() };
-                        let dpi_factor = gl_window.get_hidpi_factor();
+                        let _ = unsafe { combined_context.make_current() };
+                        let dpi_factor = combined_context.get_hidpi_factor();
                         size = logical_size.to_physical(dpi_factor);
-                        gl_window.resize(size);
+                        combined_context.resize(size);
 
                         unsafe {
-                            let _ = gl_window.make_current();
+                            let _ = combined_context.make_current();
                             glw.gl.DeleteTextures(1, &render_tex);
                             glw.gl.DeleteFramebuffers(1, &window_fb);
 
@@ -107,7 +109,7 @@ fn main() {
                                 0,
                             );
 
-                            let _ = gl_context.make_current();
+                            let _ = headless_context.make_current();
                             glc.gl.DeleteFramebuffers(1, &context_fb);
 
                             glc.gl.GenFramebuffers(1, &mut context_fb);
@@ -129,10 +131,10 @@ fn main() {
             }
         });
 
-        let _ = unsafe { gl_context.make_current() };
+        let _ = unsafe { headless_context.make_current() };
         glc.draw_frame([1.0, 0.5, 0.7, 1.0]);
 
-        let _ = unsafe { gl_window.make_current() };
+        let _ = unsafe { combined_context.make_current() };
         unsafe {
             glw.gl.BlitFramebuffer(
                 0, 0, size.width as _, size.height as _,
@@ -141,14 +143,14 @@ fn main() {
                 gl::NEAREST,
             );
         }
-        let _ = gl_window.swap_buffers();
+        let _ = combined_context.swap_buffers();
     }
 
     unsafe {
-        let _ = gl_window.make_current();
+        let _ = combined_context.make_current();
         glw.gl.DeleteTextures(1, &render_tex);
         glw.gl.DeleteFramebuffers(1, &window_fb);
-        let _ = gl_context.make_current();
+        let _ = headless_context.make_current();
         glc.gl.DeleteFramebuffers(1, &context_fb);
     }
 }